+++ docs/hooks.py
@@ -1,5 +1,6 @@
 import shutil
 
 
-def copy_history(*args, **kwargs):
+def copy_docs(*args, **kwargs):
     shutil.copy("CHANGELOG.md", "docs/changelog.md")
+    shutil.copy("README.md", "docs/index.md")