+++ tests/benchmarks/__init__.py
@@ -1,4 +1,3 @@
 # SPDX-FileCopyrightText: 2024-present David Peckham <dave.peckham@icloud.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.3.0"